requirements_repro.txt
dvc[gdrive]
nltk==3.6.7
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2
Unidecode==1.3.2
pyyaml

dvc run -n eval \
    -p eval \
    -d resources/test_data.csv \
    -d resources/model.pkl \
    -d requirements.txt \
    -d src/eval.py \
    -M metrics.json \
    python3 src/eval.py